Fix DestinationSet.removeAllExcept(Object) and add additional tests
The DestinationSet.removeAllExcept(Object) method takes a destination
metadata id or destination name. If it's passed a reference to a
destination which does not exist, then it should remove all entries from
the set. This was working correctly if a number was passed to the method
which did not refer to a destination which remained in the set, but not
for a name. This change ensures that when a destination name does not
appear in the destination set, that all destinations are removed.

Additonal tests were also added to DestinationSetTest for greater
coverage.

Signed-off-by: Tony Germano <tony@germano.name>
Issue: https://github.com/nextgenhealthcare/connect/issues/5875
Reported-by: Mike Hokanson <mhokanson@gmail.com>