Remove empty extensionmanager-client.jar
ExtensionManager apparently used to be a plugin, but it has not been for
some time. The step in the client build script that produced the
extension jar was not removed at the same time and has been producing an
effectively empty jar file with no classes and only a manifest. The
actual extensionmanager related classes are included in
mirth-client.jar.

Related: 027ae2405ee6fe66b26b29a28e5315fcaf1dace0
Signed-off-by: Tony Germano <tony@germano.name>